% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeSteps.R
\name{step_kmeans}
\alias{step_kmeans}
\title{step_kmeans: a recipe step for clustering by kmeans.}
\usage{
step_kmeans(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  num_k = NULL,
  retain = FALSE,
  id = rand_id("kmeans")
)
}
\arguments{
\item{role}{For model terms created by this step, what analysis role should
they be assigned?}

\item{trained}{A logical to indicate if the quantities for preprocessing
have been estimated.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked
by [recipes::bake.recipe()]?}

\item{num_k}{The parameter num_k is passed to the parameter centers in kmeans function.}

\item{retain}{A logical to specify whether the original predictors should be retained along with the new embedding 
variables.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
step_kmeans: a recipe step for clustering by kmeans.
}
